fix caching
some caches are not considered, because annotations on interface level are processed, iff "interface-proxies" are created. However this does not work for CGLIB proxies (proxyTargetClass=true).
My stupid quick fix is, to duplicate the annotations, to have them on both, the interface and the class level. Then either case should succeed.